Move methods up
