+++ src/main/java/com/yahoo/memory/BaseWritableMemoryImpl.java
@@ -9,11 +9,13 @@ import static com.yahoo.memory.UnsafeUtil.ARRAY_BOOLEAN_BASE_OFFSET;
 import static com.yahoo.memory.UnsafeUtil.ARRAY_BOOLEAN_INDEX_SCALE;
 import static com.yahoo.memory.UnsafeUtil.ARRAY_BYTE_BASE_OFFSET;
 import static com.yahoo.memory.UnsafeUtil.ARRAY_BYTE_INDEX_SCALE;
+import static com.yahoo.memory.UnsafeUtil.LS;
 import static com.yahoo.memory.UnsafeUtil.assertBounds;
 import static com.yahoo.memory.UnsafeUtil.checkBounds;
 import static com.yahoo.memory.UnsafeUtil.unsafe;
 
 import java.nio.ByteBuffer;
+import java.nio.ByteOrder;
 
 /*
  * Developer notes: The heavier methods, such as put/get arrays, duplicate, region, clear, fill,
@@ -251,6 +253,32 @@ abstract class BaseWritableMemoryImpl extends WritableMemory {
   @Override
   public boolean isValid() {
     return state.isValid();
+  }
+
+  @Override
+  public ByteOrder getResourceOrder() {
+    state.assertValid();
+    return state.order();
+  }
+
+  @Override
+  public boolean swapBytes() {
+    state.assertValid();
+    return state.isSwapBytes();
+  }
+
+  @Override
+  public String toHexString(final String header, final long offsetBytes, final int lengthBytes) {
+    state.checkValid();
+    final String klass = this.getClass().getSimpleName();
+    final String s1 = String.format("(..., %d, %d)", offsetBytes, lengthBytes);
+    final long hcode = hashCode() & 0XFFFFFFFFL;
+    final String call = ".toHexString" + s1 + ", hashCode: " + hcode;
+    final StringBuilder sb = new StringBuilder();
+    sb.append("### ").append(klass).append(" SUMMARY ###").append(LS);
+    sb.append("Header Comment      : ").append(header).append(LS);
+    sb.append("Call Parameters     : ").append(call);
+    return Memory.toHex(sb.toString(), offsetBytes, lengthBytes, state);
   }
 
   //PRIMITIVE putXXX() and putXXXArray() implementations XXX
